requirements.txt
player
streamlit
pytesseract
pyperclip
pillow


# streamlit
# easyocr
# Pillow
# numpy